# swagger/v1/swagger.yml
openapi: 3.0.1
info:
  title: Semantic Container API
  description: >-
    Semantic Containers enable the secure and traceable exchange of defined data
    on specified usage conditions between multiple parties.
     
     Further information:
     - learn more about Semantic Containers at https://www.ownyourdata.eu/semcon
     - view other [OwnYourData Service API's](https://api-docs.ownyourdata.eu)
  contact:
    email: office@ownyourdata.eu
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
  version: v1
servers:
  - url: /
security:
  - BearerAuth: []
paths:
...
  '/api/receipt/{id}':
    get:
      tags:
        - Data receipt
      summary: retrieve information for specified receipt (including record IDs)
      parameters:
        - name: id
          in: path
          description: hash value provided by write operation
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content: {}
        '404':
          description: not found
          content: {}
  '/api/receipt/{ttl}/{id}':
    get:
      tags:
        - Data receipt
      summary: >-
        retrieve complete information for specified receipt with maximum
        recursion level defined in ttl (time-to-live)
      parameters:
        - name: ttl
          in: path
          description: recursion level to query subsequent containers
          required: true
          schema:
            type: integer
        - name: id
          in: path
          description: hash value provided by write operation
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content: {}
        '404':
          description: not found
          content: {}
  '/api/receipt/{id}/revoke':
    delete:
      tags:
        - Data receipt
      summary: revoke all records for specified receipt
      parameters:
        - name: id
          in: path
          description: hash value provided by write operation
          required: true
          schema:
            type: string
        - name: revocation_key
          in: query
          schema:
            type: string
      responses:
        '200':
          description: success
          content: {}
        '404':
          description: not found
          content: {}
      x-codegen-request-body-name: input
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer